move keychain service to shared folder
